Bot/requirements.txt
ipykernel
python-dotenv
langchain
langchain_community
langchain_groq
langchain_chroma
langchain_huggingface
fastapi 
uvicornp
